chore: separate out create/update client
