+++ src/plugins/candyMachineModule/CandyMachineClient.ts
@@ -1,43 +1,13 @@
-import { ConfirmOptions, Keypair, PublicKey } from '@solana/web3.js';
+import { ConfirmOptions, PublicKey } from '@solana/web3.js';
+import { ModuleClient, Signer, MetaplexFile } from '@/types';
 import {
-  ModuleClient,
-  Signer,
-  convertToPublickKey,
-  MetaplexFile,
-} from '@/types';
-import {
-  CandyMachineAlreadyHasThisAuthorityError,
   CandyMachineCannotAddAmountError,
   CandyMachineIsFullError,
   CandyMachineToUpdateNotFoundError,
-  CreatedCandyMachineNotFoundError,
   UpdatedCandyMachineNotFoundError,
 } from '@/errors';
-import {
-  CandyMachineConfigWithoutStorage,
-  candyMachineDataFromConfig,
-} from './config';
-import {
-  CreateCandyMachineInput,
-  createCandyMachineOperation,
-  CreateCandyMachineOutput,
-} from './createCandyMachine';
 import { CandyMachine } from './CandyMachine';
-import {
-  UpdateCandyMachineInputWithoutCandyMachineData,
-  updateCandyMachineOperation,
-  UpdateCandyMachineOutput,
-} from './updateCandyMachine';
-import {
-  CandyMachineData,
-  ConfigLine,
-  Creator,
-} from '@metaplex-foundation/mpl-candy-machine';
-import {
-  UpdateAuthorityInput,
-  updateAuthorityOperation,
-  UpdateAuthorityOutput,
-} from './updateAuthority';
+import { ConfigLine, Creator } from '@metaplex-foundation/mpl-candy-machine';
 import {
   JsonMetadata,
   JsonMetadataFile,
@@ -50,16 +20,8 @@ import {
   findByAddress,
   findByAuthorityAndUuid,
 } from './Client.queries';
-
-export type CandyMachineInitFromConfigOpts = {
-  candyMachineSigner?: Signer;
-  authorityAddress?: PublicKey;
-  confirmOptions?: ConfirmOptions;
-};
-export type UpdateCandyMachineParams =
-  UpdateCandyMachineInputWithoutCandyMachineData & Partial<CandyMachineData>;
-
-export type UpdateCandyMachineAuthorityParams = UpdateAuthorityInput;
+import { create, createFromConfig } from './Client.create';
+import { update, updateAuthority } from './Client.update';
 
 export type AddAssetsToCandyMachineParams = {
   // Accounts
@@ -130,97 +92,14 @@ export class CandyMachineClient extends ModuleClient {
   // -----------------
   // Create
   // -----------------
-  async create(
-    input: CreateCandyMachineInput
-  ): Promise<CreateCandyMachineOutput & { candyMachine: CandyMachine }> {
-    const operation = createCandyMachineOperation(input);
-    const output = await this.metaplex.operations().execute(operation);
-
-    const candyMachine = await this.findByAddress(
-      output.candyMachineSigner.publicKey
-    );
-    if (candyMachine === null) {
-      throw new CreatedCandyMachineNotFoundError(
-        output.candyMachineSigner.publicKey
-      );
-    }
-
-    return { candyMachine, ...output };
-  }
-
-  createFromConfig(
-    config: CandyMachineConfigWithoutStorage,
-    opts: CandyMachineInitFromConfigOpts
-  ): Promise<CreateCandyMachineOutput & { candyMachine: CandyMachine }> {
-    const { candyMachineSigner = Keypair.generate() } = opts;
-    const candyMachineData = candyMachineDataFromConfig(
-      config,
-      candyMachineSigner.publicKey
-    );
-    const walletAddress = convertToPublickKey(config.solTreasuryAccount);
-
-    return this.create({
-      candyMachineSigner,
-      walletAddress,
-      authorityAddress: opts.authorityAddress,
-      ...candyMachineData,
-    });
-  }
+  create = create;
+  createFromConfig = createFromConfig;
 
   // -----------------
   // Update
   // -----------------
-  async update(
-    input: UpdateCandyMachineParams
-  ): Promise<UpdateCandyMachineOutput & { candyMachine: CandyMachine }> {
-    const currentCandyMachine = await this.findByAddress(
-      input.candyMachineAddress
-    );
-    if (currentCandyMachine == null) {
-      throw new CandyMachineToUpdateNotFoundError(input.candyMachineAddress);
-    }
-
-    const updatedData = currentCandyMachine.updatedCandyMachineData(input);
-
-    const operation = updateCandyMachineOperation({ ...input, ...updatedData });
-    const output = await this.metaplex.operations().execute(operation);
-
-    const candyMachine = await this.findByAddress(input.candyMachineAddress);
-    if (candyMachine == null) {
-      throw new UpdatedCandyMachineNotFoundError(input.candyMachineAddress);
-    }
-
-    return { candyMachine, ...output };
-  }
-
-  async updateAuthority(
-    input: UpdateCandyMachineAuthorityParams
-  ): Promise<UpdateAuthorityOutput & { candyMachine: CandyMachine }> {
-    const currentCandyMachine = await this.findByAddress(
-      input.candyMachineAddress
-    );
-    if (currentCandyMachine == null) {
-      throw new CandyMachineToUpdateNotFoundError(input.candyMachineAddress);
-    }
-
-    if (
-      currentCandyMachine.authorityAddress.equals(input.newAuthorityAddress)
-    ) {
-      throw new CandyMachineAlreadyHasThisAuthorityError(
-        input.newAuthorityAddress
-      );
-    }
-
-    const operation = updateAuthorityOperation(input);
-    const output = await this.metaplex.operations().execute(operation);
-
-    const candyMachine = await this.findByAddress(input.candyMachineAddress);
-    if (candyMachine == null) {
-      throw new UpdatedCandyMachineNotFoundError(input.candyMachineAddress);
-    }
-
-    return { candyMachine, ...output };
-  }
+  update = update;
+  updateAuthority = updateAuthority;
 
   // -----------------
   // Add and Upload Assets
